+++ src/main/java/gravit/code/chapter/service/ChapterService.java
@@ -4,11 +4,10 @@ import gravit.code.chapter.domain.ChapterRepository;
 import gravit.code.chapter.dto.response.ChapterSummary;
 import gravit.code.global.exception.domain.CustomErrorCode;
 import gravit.code.global.exception.domain.RestApiException;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
