+++ src/main/java/gravit/code/learning/controller/LearningController.java
@@ -9,6 +9,7 @@ import gravit.code.learning.controller.docs.LearningControllerDocs;
 import gravit.code.learning.dto.request.LearningSubmissionSaveRequest;
 import gravit.code.learning.dto.response.LearningSubmissionSaveResponse;
 import gravit.code.learning.facade.LearningFacade;
+import gravit.code.learning.facade.LearningQueryFacade;
 import gravit.code.lesson.dto.response.LessonDetailResponse;
 import gravit.code.lesson.dto.response.LessonResponse;
 import gravit.code.problem.dto.request.ProblemSubmissionRequest;
@@ -20,13 +21,18 @@ import gravit.code.unit.dto.response.UnitDetailResponse;
 import gravit.code.wrongAnsweredNote.dto.response.WrongAnsweredNoteDeleteRequest;
 import gravit.code.wrongAnsweredNote.service.WrongAnsweredNoteService;
 import jakarta.validation.Valid;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -34,13 +40,14 @@ import java.util.List;
 public class LearningController implements LearningControllerDocs {
 
     private final LearningFacade learningFacade;
+    private final LearningQueryFacade learningQueryFacade;
     private final ReportService reportService;
     private final BookmarkService bookmarkService;
     private final WrongAnsweredNoteService wrongAnsweredNoteService;
 
     @GetMapping("/chapters")
     public ResponseEntity<List<ChapterDetailResponse>> getAllChapters(@AuthenticationPrincipal LoginUser loginUser) {
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllChapterDetail(loginUser.getId()));
+        return ResponseEntity.status(HttpStatus.OK).body(learningQueryFacade.getAllChapterDetail(loginUser.getId()));
     }
 
     @GetMapping("/{chapterId}/units")
@@ -48,7 +55,7 @@ public class LearningController implements LearningControllerDocs {
             @AuthenticationPrincipal LoginUser loginUser,
             @PathVariable("chapterId") Long chapterId
     ) {
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllUnitDetailInChapter(loginUser.getId(), chapterId));
+        return ResponseEntity.status(HttpStatus.OK).body(learningQueryFacade.getAllUnitDetailInChapter(loginUser.getId(), chapterId));
     }
 
     @GetMapping("/{unitId}/lessons")
@@ -56,7 +63,15 @@ public class LearningController implements LearningControllerDocs {
             @AuthenticationPrincipal LoginUser loginUser,
             @PathVariable("unitId") Long unitId
     ){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllLessonInUnit(loginUser.getId(), unitId));
+        return ResponseEntity.status(HttpStatus.OK).body(learningQueryFacade.getAllLessonInUnit(loginUser.getId(), unitId));
+    }
+
+    @GetMapping("/{lessonId}")
+    public ResponseEntity<LessonResponse> getAllProblemsInLesson(
+            @AuthenticationPrincipal LoginUser loginUser,
+            @PathVariable("lessonId") Long lessonsId
+    ){
+        return ResponseEntity.status(HttpStatus.OK).body(learningQueryFacade.getAllProblemsInLesson(lessonsId, loginUser.getId()));
     }
 
     @PostMapping("/lessons/results")
@@ -76,20 +91,12 @@ public class LearningController implements LearningControllerDocs {
         return ResponseEntity.ok().build();
     }
 
-    @GetMapping("/{lessonId}")
-    public ResponseEntity<LessonResponse> getAllProblemsInLesson(
-            @AuthenticationPrincipal LoginUser loginUser,
-            @PathVariable("lessonId") Long lessonsId
-    ){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllProblemsInLesson(lessonsId, loginUser.getId()));
-    }
-
     @GetMapping("/{unitId}/bookmarks")
     public ResponseEntity<BookmarkedProblemResponse> getBookmarkedProblemsInUnit(
             @AuthenticationPrincipal LoginUser loginUser,
             @PathVariable("unitId") Long unitId
     ){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getBookmarkedProblemsInUnit(loginUser.getId(), unitId));
+        return ResponseEntity.status(HttpStatus.OK).body(learningQueryFacade.getBookmarkedProblemsInUnit(loginUser.getId(), unitId));
     }
 
     @GetMapping("/{unitId}/wrong-answered-notes")
@@ -97,7 +104,7 @@ public class LearningController implements LearningControllerDocs {
             @AuthenticationPrincipal LoginUser loginUser,
             @PathVariable("unitId") Long unitId
     ){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getWrongAnsweredProblemsInUnit(loginUser.getId(), unitId));
+        return ResponseEntity.status(HttpStatus.OK).body(learningQueryFacade.getWrongAnsweredProblemsInUnit(loginUser.getId(), unitId));
     }
 
     @DeleteMapping("/wrong-answered-notes")
