+++ src/main/java/gravit/code/unit/service/UnitService.java
@@ -4,11 +4,10 @@ import gravit.code.global.exception.domain.CustomErrorCode;
 import gravit.code.global.exception.domain.RestApiException;
 import gravit.code.unit.domain.UnitRepository;
 import gravit.code.unit.dto.response.UnitSummary;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -26,6 +25,7 @@ public class UnitService {
         return unitRepository.findUnitSummaryById(unitId)
                 .orElseThrow(() -> new RestApiException(CustomErrorCode.UNIT_NOT_FOUND));
     }
+
     @Transactional(readOnly = true)
     public UnitSummary getUnitSummaryByLessonId(long lessonId) {
         return unitRepository.findUnitSummaryByLessonId(lessonId)
